Begin refactoring to remove dependence on cookiecutter
