+++ {{ cookiecutter.repo_name }}/flowrider/__main__.py
@@ -1,0 +1,4 @@
+from flowrider.cli.cli import app
+
+if __name__ == "__main__":
+    app()
