+++ {{ cookiecutter.repo_name }}/flowrider/utils.py
@@ -11,18 +11,38 @@ from metaflow import FlowSpec
 
 logger = logging.getLogger(__file__)
 
+# GENERIC
+def load_run_id(path: Path) -> int:
+    """Load run id from `path`."""
+    return int(path.open().read())
 
+
+def is_dir(src: PathLike, x):
+    """Is `x` a directory child of `src`?."""
+    return (Path(src) / x).is_dir()
+
+
+def is_pkg_installed(pkg_name: str) -> bool:
+    """Is `pkg_name` installed in running python environment?"""
+    try:
+        subprocess.run(
+            [sys.executable, "-m", "pip", "show", pkg_name],
+            stdout=subprocess.DEVNULL,
+            stderr=subprocess.DEVNULL,
+            check=True,
+        )
+        return True
+    except subprocess.CalledProcessError:
+        return False
+
+
+# COOKIECUTTER - getter
 def run_id_path(flow_subpath: str, tag: str, src_dir: Path) -> Path:
     """Construct path to run id for flow corresponding to `{path}_{tag}`.
 
     E.g. "myflow/flow"
     """
     return src_dir / "config" / "pipeline" / f"{flow_subpath}_{tag}.run_id"
-
-
-def load_run_id(path: Path) -> int:
-    """Load run id from `path`."""
-    return int(path.open().read())
 
 
 def flow_name_from_path(flow_subpath: str, repo_name: str) -> str:
@@ -47,22 +67,3 @@ def flow_name_from_path(flow_subpath: str, repo_name: str) -> str:
 
     flow_name, _ = flows[0]
     return flow_name
-
-
-def is_dir(src: PathLike, x):
-    """Is `x` a directory child of `src`?."""
-    return (Path(src) / x).is_dir()
-
-
-def is_pkg_installed(pkg_name: str) -> bool:
-    """Is `pkg_name` installed in running python environment?"""
-    try:
-        subprocess.run(
-            [sys.executable, "-m", "pip", "show", pkg_name],
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-            check=True,
-        )
-        return True
-    except subprocess.CalledProcessError:
-        return False
