+++ {{ cookiecutter.repo_name }}/requirements.txt
@@ -9,3 +9,4 @@ tqdm
 typer
 sh
 pyyaml
+toolz
